Remove decompression from the legacy service
