+++ WordPress/Classes/Services/MediaThumbnailService.swift
@@ -2,132 +2,147 @@ import Foundation
 
 /// A service for handling the process of retrieving and generating thumbnail images
 /// for existing Media objects, whether remote or locally available.
-final class MediaThumbnailService: NSObject {
-    static let shared = MediaThumbnailService()
+///
+class MediaThumbnailService: NSObject {
+
+    /// Completion handler for a thumbnail URL.
+    ///
+    public typealias OnThumbnailURL = (URL?) -> Void
+
+    /// Error handler.
+    ///
+    public typealias OnError = (Error) -> Void
+
+    private static let defaultExportQueue: DispatchQueue = DispatchQueue(label: "org.wordpress.mediaThumbnailService", autoreleaseFrequency: .workItem)
+
+    @objc public lazy var exportQueue: DispatchQueue = {
+        return MediaThumbnailService.defaultExportQueue
+    }()
 
     private let coreDataStack: CoreDataStackSwift
 
-    init(coreDataStack: CoreDataStackSwift = ContextManager.shared) {
+    /// The initialiser for Objective-C code.
+    ///
+    /// Using `ContextManager` as the argument becuase `CoreDataStackSwift` is not accessible from Objective-C code.
+    @objc
+    convenience init(contextManager: ContextManager) {
+        self.init(coreDataStack: contextManager)
+    }
+
+    init(coreDataStack: CoreDataStackSwift) {
         self.coreDataStack = coreDataStack
     }
 
-    /// Returns a decompressed thumbnail for the given media asset.
+    /// Generate a URL to a thumbnail of the Media, if available.
     ///
     /// - Parameters:
-    ///   - media: The Media object.
-    ///   - preferredSize: An ideal size of the thumbnail in pixels. If `zero`, the maximum dimension of the UIScreen is used.
-    @MainActor
-    func image(for media: Media, preferredSize: CGSize) async throws -> UIImage {
-        guard media.remoteStatus != .stub else {
-            let media = try await fetchStubMedia(for: media)
-            // This should never happen, but adding it just in case to avoid recusion
-            guard media.remoteStatus != .stub else {
-                assertionFailure("The fetched media still has a .stub status")
-                throw MediaThumbnailExporter.ThumbnailExportError.failedToGenerateThumbnailFileURL
-            }
-            return try await image(for: media, preferredSize: preferredSize)
-        }
-
-        let imageURL = try await imageURL(for: media, preferredSize: preferredSize)
-        return try await Task.detached {
-            guard let image = UIImage(data: try Data(contentsOf: imageURL)) else {
-                throw URLError(.cannotDecodeContentData)
-            }
-            // Forces decompression (or bitmapping) to happen in the background.
-            // It's very expensive for some image formats, such as JPEG.
-            return image.preparingForDisplay() ?? image
-        }.value
-    }
-
-    /// Generate a URL to a thumbnail of the Media, saves it locally, and returns
-    /// a URL pointing to the saved file.
-    ///
-    /// - Parameters:
-    ///   - media: The Media object.
-    ///   - preferredSize: An ideal size of the thumbnail in pixels. If `zero`, the maximum dimension of the UIScreen is used.
+    ///   - media: The Media object the URL should be a thumbnail of.
+    ///   - preferredSize: An ideal size of the thumbnail in points. If `zero`, the maximum dimension of the UIScreen is used.
+    ///   - onCompletion: Completion handler passing the URL once available, or nil if unavailable. This closure is called on the `exportQueue`.
+    ///   - onError: Error handler. This closure is called on the `exportQueue`.
     ///
     /// - Note: Images may be downloaded and resized if required, avoid requesting multiple explicit preferredSizes
     ///   as several images could be downloaded, resized, and cached, if there are several variations in size.
     ///
-    @MainActor
-    func imageURL(for media: Media, preferredSize: CGSize) async throws -> URL {
-        do {
-            return try await _imageURL(for: media, preferredSize: preferredSize)
-        } catch {
-            if let error = error as? MediaExportError {
-                MediaImportService.logExportError(error)
-            }
-            throw error
-        }
-    }
-
-    @MainActor
-    private func _imageURL(for media: Media, preferredSize: CGSize) async throws -> URL {
-        let mediaID = TaggedManagedObjectID(saved: media)
-        let exporter = makeThumbnailExporter(for: media, preferredSize: preferredSize)
-
-        // Check if there is already an exported thumbnail available
-        if let identifier = media.localThumbnailIdentifier, let imageURL = await getLocalThumbnailURL(for: identifier, exporter: exporter) {
-            return imageURL
-        }
-
-        // Downloads the remote thumbnail for the asset.
-        @MainActor func getRemoteThumbnail() async throws -> URL {
-            let image = try await downloadThumbnail(forMedia: media, preferredSize: preferredSize)
-            let (identifier, export) = try await exporter.exportThumbnail(forImage: image)
-            try await saveLocalIdentifier(identifier: identifier, for: mediaID)
-            return export.url
-        }
-
-        // If the asset is available locally, export thumbnails from the local asset
-        if let localAssetURL = media.absoluteLocalURL,
-           exporter.supportsThumbnailExport(forFile: localAssetURL) {
-            let (identifier, export) = try await exporter.exportThumbnail(forFileURL: localAssetURL)
-            try await saveLocalIdentifier(identifier: identifier, for: mediaID)
-            return export.url
-        }
-
-        // If the Media item is a video and has a remote video URL, try and export from the remote video URL.
-        if media.mediaType == .video, let videoURL = media.remoteURL.flatMap(URL.init) {
+    @objc func thumbnailURL(forMedia media: Media, preferredSize: CGSize, onCompletion: @escaping OnThumbnailURL, onError: OnError?) {
+        // We can use the main context here because we only read the `Media` instance, without changing it, and all
+        // the time consuming work is done in background queues.
+        let context = coreDataStack.mainContext
+        context.perform {
+            var objectInContext: NSManagedObject?
             do {
-                let (identifier, export) = try await exporter.exportThumbnail(forVideoURL: videoURL)
-                try await saveLocalIdentifier(identifier: identifier, for: mediaID)
-                return export.url
+                objectInContext = try context.existingObject(with: media.objectID)
             } catch {
-                // If an error occurred with the remote video URL, try and download the Media's
-                // remote thumbnail instead.
-                return try await getRemoteThumbnail()
+                self.exportQueue.async {
+                    onError?(error)
+                }
+                return
             }
-        }
+            guard let mediaInContext = objectInContext as? Media else {
+                return
+            }
+            // Configure a thumbnail exporter.
+            let exporter = MediaThumbnailExporter()
+            exporter.mediaDirectoryType = .cache
+            if preferredSize == CGSize.zero {
+                // When using a zero size, default to the maximum screen dimension.
+                let screenSize = UIScreen.main.bounds
+                let screenSizeMax = max(screenSize.width, screenSize.height)
+                exporter.options.preferredSize = CGSize(width: screenSizeMax, height: screenSizeMax)
+            } else {
+                exporter.options.preferredSize = preferredSize
+            }
 
-        return try await getRemoteThumbnail()
-    }
+            // Check if there is already an exported thumbnail available.
+            if let identifier = mediaInContext.localThumbnailIdentifier, let availableThumbnail = exporter.availableThumbnail(with: identifier) {
+                self.exportQueue.async {
+                    onCompletion(availableThumbnail)
+                }
+                return
+            }
 
-    @MainActor
-    private func makeThumbnailExporter(for media: Media, preferredSize: CGSize) -> MediaThumbnailExporter {
-        let exporter = MediaThumbnailExporter()
-        exporter.mediaDirectoryType = .cache
-        if preferredSize == CGSize.zero {
-            // When using a zero size, default to the maximum screen dimension.
-            let screenSize = UIScreen.main.bounds
-            let screenSizeMax = max(screenSize.width, screenSize.height)
-            exporter.options.preferredSize = CGSize(width: screenSizeMax, height: screenSizeMax)
-        } else {
-            exporter.options.preferredSize = preferredSize
-        }
-        if let identifier = media.localThumbnailIdentifier {
-            exporter.options.identifier = identifier
-        } else {
-            exporter.options.identifier = media.objectID.uriRepresentation().lastPathComponent
-        }
-        return exporter
-    }
+            // If we already set an identifier before let's reuse it
+            if let identifier = mediaInContext.localThumbnailIdentifier {
+                exporter.options.identifier = identifier
+            } else {
+                exporter.options.identifier = media.objectID.uriRepresentation().lastPathComponent
+            }
 
-    private func getLocalThumbnailURL(for identifier: String, exporter: MediaThumbnailExporter) async -> URL? {
-        // Checking if the URL is available uses disk I/O, so moving it to background
-        await Task.detached {
-            exporter.availableThumbnail(with: identifier)
-        }.value
+            // Configure a handler for any thumbnail exports
+            let onThumbnailExport: MediaThumbnailExporter.OnThumbnailExport = { (identifier, export) in
+                self.handleThumbnailExport(media: mediaInContext,
+                                           identifier: identifier,
+                                           export: export,
+                                           onCompletion: onCompletion)
+            }
+            // Configure an error handler
+            let onThumbnailExportError: OnExportError = { (error) in
+                self.handleExportError(error, errorHandler: onError)
+            }
+
+            // Configure an attempt to download a remote thumbnail and export it as a thumbnail.
+            let attemptDownloadingThumbnail: () -> Void = {
+                self.downloadThumbnail(forMedia: mediaInContext, preferredSize: preferredSize, callbackQueue: self.exportQueue, onCompletion: { (image) in
+                    guard let image = image else {
+                        onError?(MediaThumbnailExporter.ThumbnailExportError.failedToGenerateThumbnailFileURL)
+                        return
+                    }
+                    exporter.exportThumbnail(forImage: image, onCompletion: onThumbnailExport, onError: onThumbnailExportError)
+                }, onError: { (error) in
+                    onError?(error)
+                })
+            }
+
+            // If the Media asset is available locally, export thumbnails from the local asset.
+            if let localAssetURL = mediaInContext.absoluteLocalURL,
+                exporter.supportsThumbnailExport(forFile: localAssetURL) {
+                    self.exportQueue.async {
+                        exporter.exportThumbnail(forFile: localAssetURL,
+                                                 onCompletion: onThumbnailExport,
+                                                 onError: onThumbnailExportError)
+                    }
+                return
+            }
+
+            // If the Media item is a video and has a remote video URL, try and export from the remote video URL.
+            if mediaInContext.mediaType == .video, let remoteURLStr = mediaInContext.remoteURL, let videoURL = URL(string: remoteURLStr) {
+                self.exportQueue.async {
+                    exporter.exportThumbnail(forVideoURL: videoURL,
+                                             onCompletion: onThumbnailExport,
+                                             onError: { (error) in
+                                                // If an error occurred with the remote video URL, try and download the Media's
+                                                // remote thumbnail instead.
+                                                context.perform {
+                                                    attemptDownloadingThumbnail()
+                                                }
+                    })
+                }
+                return
+            }
+
+            // Try and download a remote thumbnail, if available.
+            attemptDownloadingThumbnail()
+        }
     }
 
     /// Download a thumbnail image for a Media item, if available.
@@ -135,18 +150,19 @@ final class MediaThumbnailService: NSObject {
     /// - Parameters:
     ///   - media: The Media object.
     ///   - preferredSize: The preferred size of the image, in points, to configure remote URLs for.
-    @MainActor
-    private func downloadThumbnail(forMedia media: Media, preferredSize: CGSize) async throws -> UIImage {
-        guard let imageURL = getRemoteThumbnailURL(for: media, preferredSize: preferredSize) else {
-            throw URLError(.badURL)
-        }
-        let host = MediaHost(with: media.blog)
-        let request = try await MediaRequestAuthenticator().authenticatedRequest(for: imageURL, host: host)
-        return try await ImageDownloader.shared.image(for: request)
-    }
-
-    @MainActor
-    private func getRemoteThumbnailURL(for media: Media, preferredSize: CGSize) -> URL? {
+    ///   - callbackQueue: The queue to execute the `onCompletion` or the `onError` callback.
+    ///   - onCompletion: Completes if everything was successful, but nil if no image is available.
+    ///   - onError: An error was encountered either from the server or locally, depending on the Media object or blog.
+    ///
+    /// - Note: based on previous implementation in MediaService.m.
+    ///
+    private func downloadThumbnail(
+        forMedia media: Media,
+        preferredSize: CGSize,
+        callbackQueue: DispatchQueue,
+        onCompletion: @escaping (UIImage?) -> Void,
+        onError: @escaping (Error) -> Void
+    ) {
         var remoteURL: URL?
         // Check if the Media item is a video or image.
         if media.mediaType == .video {
@@ -167,30 +183,43 @@ final class MediaThumbnailService: NSObject {
                 }
             }
         }
-        return remoteURL
-    }
-
-    private func saveLocalIdentifier(identifier: MediaThumbnailExporter.ThumbnailIdentifier, for mediaID: TaggedManagedObjectID<Media>) async throws {
-        try await coreDataStack.performAndSave { context in
-            let media = try context.existingObject(with: mediaID)
-            if media.localThumbnailIdentifier != identifier {
-                media.localThumbnailIdentifier = identifier
+        guard let imageURL = remoteURL else {
+            // No URL's available, no images available.
+            callbackQueue.async {
+                onCompletion(nil)
             }
+            return
         }
+
+        let download = AuthenticatedImageDownload(url: imageURL, mediaHost: MediaHost(with: media.blog), callbackQueue: callbackQueue, onSuccess: onCompletion, onFailure: onError)
+
+        download.start()
     }
 
-    @MainActor
-    private func fetchStubMedia(for media: Media) async throws -> Media {
-        guard let mediaID = media.mediaID else {
-            throw MediaThumbnailExporter.ThumbnailExportError.failedToGenerateThumbnailFileURL
-        }
-        return try await withUnsafeThrowingContinuation { continuation in
-            let mediaService = MediaService(managedObjectContext: coreDataStack.mainContext)
-            mediaService.getMediaWithID(mediaID, in: media.blog, success: {
-                continuation.resume(returning: $0)
-            }, failure: {
-                continuation.resume(throwing: $0)
-            })
+    // MARK: - Helpers
+
+    private func handleThumbnailExport(media: Media, identifier: MediaThumbnailExporter.ThumbnailIdentifier, export: MediaExport, onCompletion: @escaping OnThumbnailURL) {
+        coreDataStack.performAndSave({ context in
+            let object = try context.existingObject(with: media.objectID)
+            // It's safe to force-unwrap here, since the `object`, if exists, must be a `Media` type.
+            let mediaInContext = object as! Media
+            mediaInContext.localThumbnailIdentifier = identifier
+        }, completion: { (result: Result<Void, Error>) in
+            switch result {
+            case .success:
+                onCompletion(export.url)
+            case .failure:
+                onCompletion(nil)
+            }
+        }, on: exportQueue)
+    }
+
+    /// Handle the OnError callback and logging any errors encountered.
+    ///
+    private func handleExportError(_ error: MediaExportError, errorHandler: OnError?) {
+        MediaImportService.logExportError(error)
+        exportQueue.async {
+            errorHandler?(error.toNSError())
         }
     }
 }
